+++ src-tauri/src/main.rs
@@ -3,86 +3,26 @@
     windows_subsystem = "windows"
 )]
 
+mod commands;
 mod config;
 mod db;
+mod logging;
+mod tray;
 
-use log::LevelFilter;
 use std::env;
-use std::fs::OpenOptions;
-use std::path::PathBuf;
-use tauri::{
-    CustomMenuItem, GlobalShortcutManager, Manager, SystemTray, SystemTrayEvent, SystemTrayMenu,
-};
+use tauri::{GlobalShortcutManager, Manager, SystemTrayEvent};
 use tauri_plugin_positioner::{Position, WindowExt};
 use tokio::runtime::Runtime;
 
-use std::sync::atomic::{AtomicBool, Ordering};
-use tauri::State;
+use std::sync::atomic::Ordering;
 
-struct SaveDialogState {
-    pub is_open: AtomicBool,
-}
-
-impl Default for SaveDialogState {
-    fn default() -> Self {
-        SaveDialogState {
-            is_open: AtomicBool::new(false),
-        }
-    }
-}
-
-#[tauri::command]
-fn open_save_dialog(state: State<SaveDialogState>) {
-    state.is_open.store(true, Ordering::SeqCst);
-}
-
-#[tauri::command]
-fn close_save_dialog(state: State<SaveDialogState>) {
-    state.is_open.store(false, Ordering::SeqCst);
-}
-
-fn get_log_path() -> PathBuf {
-    let home_dir = dirs::home_dir().expect("Could not find the home directory");
-    home_dir.join(".kftray").join("app.log")
-}
-
-fn setup_logging() {
-    let log_filter = match env::var("RUST_LOG") {
-        Ok(filter) => filter.parse().unwrap_or(LevelFilter::Info),
-        Err(_) => LevelFilter::Off,
-    };
-
-    if env::var("KFTRAY_DEBUG").is_ok() {
-        let log_path = get_log_path();
-        let log_dir = log_path.parent().expect("Could not find the log directory");
-        std::fs::create_dir_all(log_dir).expect("Could not create log directory");
-
-        let log_file = OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(log_path)
-            .expect("Could not open log file");
-
-        env_logger::Builder::from_default_env()
-            .filter_level(log_filter)
-            .format_timestamp_secs()
-            .target(env_logger::Target::Pipe(Box::new(log_file)))
-            .init();
-    } else {
-        env_logger::Builder::new()
-            .filter_level(log_filter)
-            .format_timestamp_secs()
-            .init();
-    }
-}
+use commands::SaveDialogState;
 
 fn main() {
-    setup_logging();
+    logging::setup_logging();
     let _ = fix_path_env::fix();
-    let quit = CustomMenuItem::new("quit".to_string(), "Quit").accelerator("CmdOrCtrl+Shift+Q");
-    let open = CustomMenuItem::new("open".to_string(), "Open App");
-    let system_tray_menu = SystemTrayMenu::new().add_item(open).add_item(quit);
 
+    let system_tray = tray::create_tray_menu();
     tauri::Builder::default()
         .manage(SaveDialogState::default())
         .setup(move |app| {
@@ -120,7 +60,7 @@ fn main() {
             Ok(())
         })
         .plugin(tauri_plugin_positioner::init())
-        .system_tray(SystemTray::new().with_menu(system_tray_menu))
+        .system_tray(system_tray)
         .on_system_tray_event(|app, event| {
             tauri_plugin_positioner::on_tray_event(app, &event);
             match event {
@@ -224,8 +164,8 @@ fn main() {
             config::update_config,
             config::export_configs,
             config::import_configs,
-            open_save_dialog,
-            close_save_dialog,
+            commands::open_save_dialog,
+            commands::close_save_dialog,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
