+++ src/assets/theme.tsx
@@ -1,4 +1,3 @@
-// theme.tsx
 import { extendTheme, ThemeConfig } from '@chakra-ui/react'
 
 const config: ThemeConfig = {
